requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
coverage==4.5.4
coveralls==1.8.2
docopt==0.6.2
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==7.2.0
nose==1.3.7
nose2==0.9.1
packaging==19.2
pluggy==0.13.0
psycopg2==2.8.4
py==1.8.0
pyparsing==2.4.3
pytest==5.2.2
pytest-cov==2.8.1
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
